E2E tests to 1 file
