+++ test/e2e/fileformatting_test.go
@@ -19,7 +19,7 @@ import (
 	"sigs.k8s.io/secrets-store-csi-driver/apis/v1alpha1"
 )
 
-type resultJson struct {
+type resultObject struct {
 	Alias  string `json:"SECRET_1"`
 	Secret string `json:"secret1"`
 }
@@ -31,7 +31,7 @@ var _ = Describe("When deploying SecretProviderClass CRD with secrets and filefo
 		ns                   *corev1.Namespace
 		nodePublishSecretRef *corev1.Secret
 		p                    *corev1.Pod
-		result               resultJson
+		result               resultObject
 	)
 
 	BeforeEach(func() {
@@ -127,6 +127,113 @@ var _ = Describe("When deploying SecretProviderClass CRD with secrets and filefo
 	})
 
 	It("should validate secret value from json file in pod", func() {
+		Expect(result.Secret).To(Equal(config.SecretValue))
+	})
+})
+
+var _ = Describe("When deploying SecretProviderClass CRD with secrets and fileformatting is `Yaml`", func() {
+	var (
+		specName             = "secret"
+		spc                  *v1alpha1.SecretProviderClass
+		ns                   *corev1.Namespace
+		nodePublishSecretRef *corev1.Secret
+		p                    *corev1.Pod
+		result               resultObject
+	)
+
+	BeforeEach(func() {
+		ns = namespace.Create(namespace.CreateInput{
+			Creator: kubeClient,
+			Name:    specName,
+		})
+
+		nodePublishSecretRef = secret.Create(secret.CreateInput{
+			Creator:   kubeClient,
+			Name:      "secrets-store-creds",
+			Namespace: ns.Name,
+			Data:      map[string][]byte{"clientid": []byte(config.AzureClientID), "clientsecret": []byte(config.AzureClientSecret)},
+		})
+
+		keyVaultObjects := []provider.KeyVaultObject{
+			{
+				ObjectName: "secret1",
+				ObjectType: provider.VaultObjectTypeSecret,
+			},
+			{
+				ObjectName:  "secret1",
+				ObjectType:  provider.VaultObjectTypeSecret,
+				ObjectAlias: "SECRET_1",
+			},
+		}
+
+		yamlArray := provider.StringArray{Array: []string{}}
+		for _, object := range keyVaultObjects {
+			obj, err := yaml.Marshal(object)
+			Expect(err).To(BeNil())
+			yamlArray.Array = append(yamlArray.Array, string(obj))
+		}
+
+		objects, err := yaml.Marshal(yamlArray)
+		Expect(err).To(BeNil())
+
+		spc = secretproviderclass.Create(secretproviderclass.CreateInput{
+			Creator:   kubeClient,
+			Config:    config,
+			Name:      "azure",
+			Namespace: ns.Name,
+			Spec: v1alpha1.SecretProviderClassSpec{
+				Provider: "azure",
+				Parameters: map[string]string{
+					"keyvaultName":   config.KeyvaultName,
+					"tenantId":       config.TenantID,
+					"fileFormatting": "Yaml",
+					"objects":        string(objects),
+				},
+			},
+		})
+
+		p = pod.Create(pod.CreateInput{
+			Creator:                  kubeClient,
+			Config:                   config,
+			Name:                     "nginx-secrets-store-inline-crd-yaml",
+			Namespace:                ns.Name,
+			SecretProviderClassName:  spc.Name,
+			NodePublishSecretRefName: nodePublishSecretRef.Name,
+		})
+
+	})
+
+	AfterEach(func() {
+		Cleanup(CleanupInput{
+			Namespace: ns,
+			Getter:    kubeClient,
+			Lister:    kubeClient,
+			Deleter:   kubeClient,
+		})
+	})
+
+	It("should parse secrets as yaml file in pod", func() {
+		pod.WaitFor(pod.WaitForInput{
+			Getter:         kubeClient,
+			KubeconfigPath: kubeconfigPath,
+			Config:         config,
+			PodName:        p.Name,
+			Namespace:      ns.Name,
+		})
+
+		cmd := getPodExecCommand("cat /mnt/secrets-store/secrets.yaml")
+		secretsYaml, err := exec.KubectlExec(kubeconfigPath, p.Name, p.Namespace, strings.Split(cmd, " "))
+		Expect(err).To(BeNil())
+
+		err = yaml.Unmarshal([]byte(secretsYaml), &result)
+		Expect(err).To(BeNil())
+	})
+
+	It("should validate alias value from yaml file in pod", func() {
+		Expect(result.Alias).To(Equal(config.SecretValue))
+	})
+
+	It("should validate secret value from yaml file in pod", func() {
 		Expect(result.Secret).To(Equal(config.SecretValue))
 	})
 })
